# openapi.yaml
openapi: 3.1.3
info:
  title: ApViMad API REST
  description: |-
    Esta API se encarga de manejar la información de la base de datos de la empresa ApViMad.
  version: 1.0.11
servers:
  - url: https://localhost/api/v3
paths:
  /viviendas:
    get:
      summary: GET de las viviendas
      description: Muestra todas las viviendas disponibles
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/viviendas"

    post:
      summary: Crea una nueva vivienda
      description: Para crear una nueva vivienda cumplir y rellenar los campos descritos en el schema de vivienda asi como el tipo de dato de los mismos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/vivienda"
      responses:
        "201":
          description: "Created"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/nuevaVivienda"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
  /viviendas/{id}:
    parameters:
      - $ref: "#/components/parameters/id"
    get:
      summary: Obtener información de una vivienda concreta
      description: Nos permite obtener y visualzar los datos de una vivienda concreta
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/vivienda"
    put:
      summary: Editar una vivienda
      description: Permite modificar los datos de una vivienda concreta siempre y cuando se cumplan los requisitos de su schema
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/actualizar_vivienda"
      responses:
        "200":
          description: "Ok"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/nuevaVivienda"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
    delete:
      summary: Eliminar una vivienda
      description: Borra una vivienda de forma definitiva
      responses:
        "200":
          description: "Ok"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"

  /alquiler:
    get:
      summary: GET de alquiler
      description: Muestra todas los alquileres disponibles
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/alquiler"

      post:
        summary: Crea una nueva vivienda de alquiler
        description: Para crear una nueva vivienda cumplir y rellenar los campos descritos en el schema de vivienda asi como el tipo de dato de los mismos
        requestBody:
          required: true
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/alquiler"
        responses:
          "201":
            description: "Created"
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/nuevoAlquiler"
          "404":
            $ref: "#/components/responses/NotFound"
          "500":
            $ref: "#/components/responses/ServerError"
  /alquiler/{id}:
    parameters:
      - $ref: "#/components/parameters/id"
    get:
      summary: Obtener información de un alquiler concreto
      description: Permite al usuario obtener y visualzar los datos de las viviendas en alquiler
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/alquiler"
    put:
      summary: Editar un alquiler
      description: Permite al usuario modificar un alquiler
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/actualizar_alquiler"
      responses:
        "200":
          description: "Ok"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/nuevoUsuario"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
    delete:
      summary: Eliminar un alquiler
      description: Permite al usuario borrar un alquiler
      responses:
        "200":
          description: "Ok"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  parameters:
    id:
      description: identificador unico de la vivienda o el usuario
      name: id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/id"
  responses:
    BadRequest:
      description: (BadRequest) los datos enviados son incorrectos o hay datos obligatorios no enviados
    Unauthorized:
      description: (Unauthorized) las credenciales de autorización no son validas
    NotFound:
      description: (NotFound) no se encontró la información
    ServerError:
      description: Error en el servidor
  schemas:
    id:
      type: integer
      minimum: 1
    viviendas:
      type: array
      minItems: 1
      items:
        $ref: "#/components/schemas/vivienda"
    vivienda:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/id"
        descripcion:
          type: string
          description: resumen de la vivienda
        barrio:
          type: string
          description: barrio donde se encuentra la vivienda
        calle:
          type: string
          description: calle donde se encuentra la vivienda
        numero:
          type: integer
          description: numero de portal de la vivienda
        piso:
          type: string
          description: planta y puerta de la vivienda
        propietario:
          type: string
          description: dueño de la vivienda
        precio:
          type: string
          description: coste de la vivienda
    nuevaVivienda:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/id"
        url:
          description: Enlace a la vivienda creada
          type: string
          format: uri
    actualizar_vivienda:
      required:
        - id
        - descripcion
        - barrio
        - calle
        - numero
        - piso
        - propietario
        - precio
      $ref: "#/components/schemas/vivienda"
      url:
        description: Enlace a la vivienda editada
        type: string
        format: uri

    alquiler:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/id"
        descripcion:
          type: string
          description: resumen de la vivienda en alquiler
        barrio:
          type: string
          description: barrio donde se encuentra la vivienda en alquiler
        calle:
          type: string
          description: calle donde se encuentra la vivienda en alquiler
        numero:
          type: integer
          description: numero de portal de la vivienda en alquiler
        piso:
          type: string
          description: planta y puerta de la vivienda en alquiler
        propietario:
          type: string
          description: dueño de la vivienda en alquiler
        precio:
          type: string
          description: coste de la vivienda en alquiler
    nuevoAlquiler:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/id"
        url:
          description: Enlace al alquiler creada
          type: string
          format: uri
    actualizar_alquiler:
      required:
        - id
        - descripcion
        - barrio
        - calle
        - numero
        - piso
        - propietario
        - precio
      $ref: "#/components/schemas/vivienda"
      url:
        description: Enlace al alquiler editado
        type: string
        format: uri